requirements.txt
langchain==0.1.4
PyPDF2==3.0.1
streamlit==1.30.0
python-dotenv
openai
tiktoken
langchain-openai
flask
flask_cors







